requirements.txt
lightning==2.3.3
matplotlib==3.9.1
numpy==2.0.0
pandas==2.2.2
Pillow==10.4.0
scikit_learn==1.5.1
torch==2.3.1+cu118
torchinfo==1.8.0
torchmetrics==1.4.0.post0
torchvision==0.18.1+cu118
tqdm==4.66.1
